pyproject.toml
[tool.poetry]
name = "cottage_hearth"
version = "0.1.0"
description = "Cottage Labs' website and internal site"
authors = ["Cottage Labs <us@cottagelabs.com>"]

[tool.poetry.dependencies]
python = "^3.9"
wagtail = "^2.13"
django-debug-toolbar = "^3.2.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
